Using public static property in method instead of ParseObjectWrapper as a means to pass information from view to view.
I put a lot of effort into following the ParseObjectWrapper pattern,
but it didn't work. I think the main reason for it not working was
because I use the saveEventually API, so the object isn't guaranteed to
have an id, which I think causes it  to not be passed via the intents.